backend/requirements.txt
# ==== Core Backend ====
Flask>=3.0.0
Flask-Cors>=4.0.1
gunicorn>=21.0.0

# ==== AI & NLP ====
google-generativeai>=0.8.0

# ==== Presentation Generation ====
python-pptx>=0.6.0
Pillow>=9.0.0

# ==== Voice Generation ====
gTTS>=2.3.0

# ==== Video Creation ====
moviepy>=1.0.0
imageio>=2.25.0
imageio-ffmpeg>=0.4.0

# ==== Utilities ====
requests>=2.31.0
numpy>=1.24.0
tqdm>=4.65.0
Werkzeug>=3.0.0
pypdf>=4.0.0
pymongo>=4.6.0
certifi>=2024.0.0
PyJWT>=2.8.0

# ==== Cloud Storage ====
cloudinary>=1.36.0

# ==== Optional (for development/debug) ====
python-dotenv>=1.0.0
